Implement Factory for all Algebras
Bite the bullet and do it manually. Now, we finally only parse once
and use the factory to later visit the tree.

Deciding to stick with this approach instead of doing the Future
conversion, which ended up resulting in ugly and likely slow
code. So I kept the phase tracking but will have to manually
ensure that values in the database are available (aka execute phases
in the correct order).